Add missing mutex lock
The functions *Sentinel.putToTop and *Sentinel.putToBottom were meant
to be called with the *Sentinel.mu mutex held. However, putToTop was
being called _without_ holding the mutex. This commit moves the
locking/unlocking into the helper functions so that the caller cannot
forget to lock.

While we are here, we simplify the helper functions.